+++ .idea/workspace.xml
@@ -89,6 +89,9 @@
       <recent name="C:\2024-1_App1\client\front\assets\images" />
     </key>
     <key name="MoveFile.RECENT_KEYS">
+      <recent name="C:\Users\user\Desktop\wap\client\front\lib\main\errand_list\check_errand" />
+      <recent name="C:\Users\user\Desktop\wap\client\front\lib\main\errand_list" />
+      <recent name="C:\Users\user\Desktop\wap\client\front\lib\main" />
       <recent name="C:\2024-1_App1\client\front\assets\images" />
     </key>
   </component>
