+++ client/front/lib/main/errand_list/errand_list.dart
@@ -10,9 +10,9 @@ import 'package:intl/intl.dart';
 import 'package:http/http.dart' as http;
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'dart:convert';
-import 'check_errand.dart';
-import 'login.dart';
-import 'status_client.dart';
+import 'check_errand/check_errand.dart';
+import '../../login.dart';
+import '../../status_client.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
