+++ client/front/lib/map.dart
@@ -7,7 +7,7 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:http/http.dart' as http;
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
-import 'errand_list.dart';
+import 'main/errand_list/errand_list.dart';
 import 'write_errand.dart';
 final storage = FlutterSecureStorage();
 void _insertOverlay(BuildContext context) {
