+++ client/front/lib/show_errand/show_errand_widget/stamp/namelength/name_length2.dart
