+++ client/front/lib/show_errand/show_errand_widget/stamp/stamp.dart
