+++ client/front/lib/sign_up/sign_up_complete.dart
@@ -8,7 +8,7 @@ import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_svg/svg.dart';
 import 'dart:async';
 import '../login.dart';
-import '../errand_list.dart';
+import '../main/errand_list/errand_list.dart';
 import 'package:http/http.dart' as http;
 import 'package:flutter_screenutil/flutter_screenutil.dart';
