+++ client/front/lib/write_errand.dart
@@ -13,8 +13,8 @@ import 'package:geolocator/geolocator.dart';
 import '../find_destination.dart';
 import 'package:http/http.dart' as http;
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
-import 'check_errand.dart';
-import 'errand_list.dart';
+import 'main/errand_list/check_errand/check_errand.dart';
+import 'main/errand_list/errand_list.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 final storage = FlutterSecureStorage();
 void _insertOverlay(BuildContext context) {
